move utils out of view.py
